pyproject.toml
[project]
name = "platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "f90nml>=1.4.5",
    "h5py>=3.14.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pexpect>=4.9.0",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
]
